refactor special configs; group list configs and function configs
